* switched more uint32 to off_t for 64bit support * added error messages in InodeAllocator, Inode * if BlockAllocator can't initialize, don't fail completely but switch to readonly * fixed a bug in FindNextMarked() for bitmaps with a length non multiple of 32 * Inode::FindBlock() now returns an optional block_run length, useful for get_file_map() * added flag for Inode for extents
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@39234 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/ext2/DirectoryIterator.cpp
@@ -502,7 +502,7 @@ DirectoryIterator::_SplitIndexedBlock(Transaction& transaction,
 		return B_NO_MEMORY;
 	ArrayDeleter<uint8> bufferDeleter(buffer);
 
-	uint32 firstPhysicalBlock = 0;
+	off_t firstPhysicalBlock = 0;
 
 	// Prepare block to hold the first half of the entries and fill the buffer
 	CachedBlock cachedFirst(fVolume);
